+++ lib/src/models/pdf_body_class.dart
@@ -3,7 +3,7 @@ import 'package:pdf/widgets.dart' as pw;
 import 'package:pdf/widgets.dart';
 import '../../greytrix_ui_kit.dart';
 
-class PdfBody{
+class PdfBody {
   PdfBody({
     this.maxRow,
     this.maxColumn,
@@ -14,7 +14,7 @@ class PdfBody{
   final List<PdfBodyField> pdfBodyFields;
 }
 
-class PdfBodyField{
+class PdfBodyField {
   PdfBodyField({
     this.row,
     this.column,
@@ -30,9 +30,8 @@ class PdfBodyField{
     this.displayKey = false,
     this.tableCellAlignment = pw.Alignment.center,
     this.tableBodyDecoration = const pw.BoxDecoration(
-      borderRadius: const pw.BorderRadius.all(pw.Radius.circular(2)),
-      color:  PdfColors.teal
-    ),
+        borderRadius: const pw.BorderRadius.all(pw.Radius.circular(2)),
+        color: PdfColors.teal),
     this.tableBodyHeight = 25,
     this.tableCellHeight = 20,
     this.tableCellAlignments = const {
@@ -42,26 +41,30 @@ class PdfBodyField{
       3: pw.Alignment.center,
       4: pw.Alignment.center,
     },
-    this.tableBodyStyle = const pw.TextStyle(fontSize: 10,),
-    this.tableCellStyle = const pw.TextStyle(fontSize: 10,),
+    this.tableBodyStyle = const pw.TextStyle(
+      fontSize: 10,
+    ),
+    this.tableCellStyle = const pw.TextStyle(
+      fontSize: 10,
+    ),
     this.tableRowDecoration = const pw.BoxDecoration(
-                border: pw.Border(
-                  bottom: pw.BorderSide(
-                    color:  PdfColors.blueGrey900,
-                    width: .5,
-                  ),
-                ),
-              ),
-    this.lineDecoration = const pw.BoxDecoration(
-          border: pw.Border(top: pw.BorderSide(
-            )),
+      border: pw.Border(
+        bottom: pw.BorderSide(
+          color: PdfColors.blueGrey900,
+          width: .5,
         ),
-        this.cellDecoration,
-        this.key,
-        this.bodyFieldKeyValueFormat = GtBodyFieldKeyValueFormat.COLUMN,
-        this.decoration,
-        this.mainAxisAlignment = pw.MainAxisAlignment.start,
-        this.crossAxisAlignment = pw.CrossAxisAlignment.start,
+      ),
+    ),
+    this.lineDecoration = const pw.BoxDecoration(
+      border: pw.Border(top: pw.BorderSide()),
+    ),
+    this.cellDecoration,
+    this.key,
+    this.bodyFieldKeyValueFormat = GtBodyFieldKeyValueFormat.COLUMN,
+    this.decoration,
+    this.mainAxisAlignment = pw.MainAxisAlignment.start,
+    this.crossAxisAlignment = pw.CrossAxisAlignment.start,
+    this.isSpanningWidget = false,
   });
 
   final int row;
@@ -76,6 +79,7 @@ class PdfBodyField{
   final pw.TextStyle valueTextStyle;
   final pw.TextStyle keyTextStyle;
   final bool displayKey;
+
   /// TABLE UI INPUTS
   final pw.Alignment tableCellAlignment;
   final pw.BoxDecoration tableBodyDecoration;
@@ -92,7 +96,13 @@ class PdfBodyField{
   final pw.BoxDecoration decoration;
   final pw.MainAxisAlignment mainAxisAlignment;
   final pw.CrossAxisAlignment crossAxisAlignment;
+
+  ///USED FOR SPANNING BODY FIELDS WIDGETS LIKE PDF_TABLE
+  ///USED TO SOLVE PDF TABLE ISSUE WHICH WAS CAUSING PROBLEM IF ONE CELL CONTENT IS TOO LARGE TO FIT IN ONE PDF PAGE
+  ///IT DOESN't ADD ANY PARENT WIDGETS TO THE PDF TABLE WIDGET AND LET IT SPAN
+  final bool isSpanningWidget;
 }
+
 enum GTBodyFieldType { ICON, TITLE, SUBTITLE, TABLE, LINE }
 
-enum GtBodyFieldKeyValueFormat {ROW, COLUMN}
+enum GtBodyFieldKeyValueFormat { ROW, COLUMN }