+++ lib/src/widgets/index.dart
@@ -41,5 +41,5 @@ export './chatbot/chatbot.dart';
 export 'gt_search_dialog.dart';
 export 'gt_app_button.dart';
 export './pdf_widget/pdf_widget.dart';
-export 'dropdown_search.dart';
+export 'gt_dropdown_search.dart';
 export 'back_drop_List_view_widget.dart';